feat(v7/core): Deprecate `Hub` class (#11528)
Deprecate the `Hub` class which we'll remove in a follow-up PR
in v8. The replacement is to mostly use the scope-centered APIs. Most
APIs that relied on the `Hub` class (unfortunately quite a lot - looks
like we weren't really consistent in using the Hub interface instead of
the class) are already deprecated and already removed in v8.