feat(core): Remove most Hub class exports (#11536)
Remove top-level exports of the `Hub` class from all SDK
packages. The only package that still exports it is core for now because
we temporarily still need the class until we shimmed the return value of
`getCurrentHub`.

Furthermore also deprecate the `Hub` interface. However, this
interface will not be removed in v8 (prob v9) to keep support for
`getCurrentHub` during the v8 life cycle.