feat(core): Remove `Hub` class export (#11560)
Remove the last public export of the `Hub` class. Now that it's
gone there are no more breaking changes (🤞). The class still remains in
the package because it's still used internally for now but we can remove
it (and shave off some well needed bytes) throught the v8 beta cycle.