feat(core): Extract errors from props in unkown inputs (#11526)
This changes the handling of unknown inputs for e.g.
`captureException()`, to check if there is a property that holds an
`Error` and use this.

So e.g. this:

```js
const error = new Error('Some error');
Sentry.captureException({ message: 'something happened!', err: error });
```

Will use `Some error` as the error message, and also take the stacktrace
from the error instead of the syntethic one.

There is of course the chance that this leads to false positives, if
something is captured and contains a reference to an unrelated error
instance. But I'd say this is not so likely, and we still keep the
serialised object as `extra` on the event.

Should fix e.g.
https://peated.sentry.io/issues/5166150355/?project=4505138086019073