feat(node): Ensure tracing without performance (TWP) works (#11564)
This fixes the node SDK to make tracing without performance work as
expected.

This has three aspects:

1. Ensure we always have an url to check against
`tracePropagationTargets`, even if the span is not sampled.
2. Ensure we use the correct propagation context always
3. Ensure we actually consistently handle ougoing http.client spans
without an active transaction

The fix for 2. was to have a fallback behavior if we don't have a scope
linked to a context yet (we'll just use the current scope then). This is
needed because we won't necessarily have an assigned scope yet at the
time this is called, because at this point in time this is still a
floating context. This also meant removing the fallback code to look at
the remote span context (because we'll always have a scope now).

The fix for 3. was to move this out of the http integration and do this
in the sampler instead.

The fix for 1. was a bit more elaborate, because we require different
things to solve the http instrumentation and the fetch instrumentation.

For fetch, we leverage our sampler, which can attach trace state that
even unsampled spans will have. We attach the URL as trace state which
we can read in the propagator, ensuring this works even for unsampled
spans.

Sadly, for http this is not sufficient, because our usage of
`onlyIfParentForOutgoingRequest` lead to some spans being created
outside of the sampler. By removing this option (see 3) and doing stuff
in the sampler instead, we can ensure that all spans for http are
handled properly.

This also bumps our node-fetch otel instrumentation to
[1.2.0](https://github.com/gas-buddy/opentelemetry-instrumentation-fetch-node/releases/tag/v1.2.0)
which fixes headers for Node 21 (tests were actually failing without
this!).