feat(node): Collect Local Variables via a worker (#11586)
Our Local Variables integration is limited by the fact that it executes
debugger commands from the main app thread. This means that no async
work can be completed while the debugger is paused and local variables
are being collected. This in turn means that all the synchronous work
needs to be queued which risks stack overflows. For this reason we've
limited local variable collection to the 5 frames closest to the top of
the stack.

When debugging from a worker thread, there are no such limitations
around performing async work during debugger pause.

This PR adds a worker thread which is tasked with collecting local
variables when the debugger pauses. It passes these back to the main
thread via IPC which are then later added to event stack frames.

## Possible Improvements

It's possible to combine both ANR and Local Variables into a single
worker thread.