fix(node): Ensure DSC is correctly set in envelope headers (#11628)
This was not correctly set before when we had an active span - the
`error-active-span` & `error-active-span-unsampled` scenarios failed
before and had no `trace` set on the envelope headers.

The fix was to actually fix our `setupEventContextTrace` utility for
OTEL, where we now just set the dsc on the event sdkProcessingMetadata,
which is picked up when event processing. I also moved this to
`preprocessEvent` to ensure this runs before other stuff.

This should fix
https://github.com/getsentry/sentry-javascript-examples/pull/7/files#r1561220216
which @s1gr1d found in her example app. 🚀