feat(opentelemetry): Update OTEL packages & relax some version ranges (#11580)
This bumps all our OTEL dependencies to require the most up to date
versions as of today for all the core packages.

This allows us to use the new semantic attributes, which esp. also some
instrumentation uses in newer versions. By requiring this as a minimum
version, we can ensure that we can update all our instrumentation.

This also relaxes all of the core packages to `^` range, so users can
easier use `@sentry/node` together with their own otel instrumentation.
The instrumentation we add remains hard-pinned.

To avoid deprecation warnings I updated all the semantic attributes
usage to the new syntax.