feat(browser): Update `propagationContext` on `spanEnd` to keep trace consistent (#11631)
As spec'd out in #11599 and agreed upon in internal discussions, a trace
should to stay consistent over the entire time span of one route.
Therefore, when the initial pageload or navigation span ends, we update
the scope's propagation context to keep span-specific attributes like
the `sampled` decision and the dynamic sampling context on the scope's
propagation context, even after the transaction has ended.

This ensures that the trace data is consistent for the entire duration
of the route. Subsequent navigations will reset the propagation context
(see #11377).