test(ci): Adjust `detectFlakyTests` to account for multiple tests in a file (#11653)
Add a test detection heuristic to the detector to roughly
count the number of tests in a file. This should reduce the number of
times each test is running in case files with multiple tests have
changed. Furthermore, this PR also refactors the run count determination
logic a bit to make it easier to understand.