feat(opentelemetry): Add addOpenTelemetryInstrumentation (#11667)
Adds `addOpenTelemetryInstrumentation`, a helper that can be used to
dynamically register OpenTelemetry instrumentation. This helps unblock
work on https://github.com/getsentry/sentry-javascript/pull/11548

The helper itself is quite small:

```ts
export function addOpenTelemetryInstrumentation(
  instrumentation: InstrumentationOption | InstrumentationOption[],
): void {
  registerInstrumentations({
    instrumentations: Array.isArray(instrumentation) ? instrumentation : [instrumentation],
  });
}
```

and is designed to accept either a standalone instrumentation or an
array of instrumentations. This gives users a ton of flexibility into
usage!