fix(angular): run tracing stuff outside Angular
This commit updates all tracing functionality to run outside the Angular zone.
Before this change, it hindered server-side rendering and hydration, causing instability
in the app. The app achieves stability when there are no micro/macro tasks running. As a
result, the HTML can now be serialized and sent to the client.