feat(core): Increase default transport buffer size from 30 to 64
Nowadays, the SDK simply sends more data and we are seeing overflows in our E2E tests which do not produce an unreasonable amount of events.

Dropped events can lead to broken traces and confusing data inside Sentry, which is why I think bumping this is reasonable.

We make this under the assumption that an event in the transport holds at max something like 5MB of memory, which would bump the memory consumption from 150MB to 320MB.