feat(core): Set default scope for BaseClient methods (#11775)
`client.captureException/captureMessage/captureEvent` do not include
current scope in events by default. This is in contrast to
`captureException/captureMessage/captureEvent` exported from every SDK
which include the current scope by default.

This PR changes the parameter names (`scope` > `currentScope`) and adds
jsdocs to make this more clear. It also changes the parameters to use
the scope interface instead of the scope class.