feat(fedback): Convert CDN bundles to use async feedback for lower bundle sizes (#11791)
It's more aggressive to start with the async loading strategy in the CDN
bundle, but could be safer because if we want to change from async->sync
it would not really be a majr version change. Whereas going from
sync->async would be more of a big feature, riskier, and could demand a
version bump or something.