fix(node): Fix nest.js error handler (#11874)
This was brought up here
https://github.com/getsentry/sentry-javascript/discussions/5578#discussioncomment-9284317,

our error handler implementation was too naive, and our tests not ideal
- the tests only checked that stuff is sent to sentry (which it was!)
but not that the page otherwise worked.

Now, I updated the test to ensure this works as expected.

With this PR, the signature for `Sentry.setupNestErrorHandler()` changes
( breaking change, but sadly required at this point). You have to pass
in an exception filter, which we'll extend to _also_ send exceptions to
Sentry:


```js
import { BaseExceptionFilter, HttpAdapterHost } from '@nestjs/core';

const { httpAdapter } = app1.get(HttpAdapterHost);
Sentry.setupNestErrorHandler(app1, new BaseExceptionFilter(httpAdapter));
```

This is a bit more involved, but also allows you to use a custom filter
if needed (we can extend _any_ exception filter).