feat(feedback): Add `captureFeedback` method (#11428)
This PR adds a new `captureFeedback` method which is exported from all
SDKs.

This method can be used to capture a (new) user feedback from any
package.

We follow the same semantics as for other capture methods like
`captureException()` or `captureMessage()`: The method returns a string,
which is the event id. We do not wait for sending to be successful or
not, we just try to send it and return. You can both set an
`associatedEventId` (which replaces the event_id of the "old"
`captureUserFeedback`), and also pass attachments to send along (which
for now are sent as a separate envelope).

For usage in the modal UI, there is still `sendFeedback` which continues
to return a promise that resolves with the event ID, or rejects with a
meaningful error message if sending fails.

This also deprecates `captureUserFeedback()`, which is only exported in
browser. We cannot remove this yet because `captureFeedback` will only
work on newer self-hosted instances, so not all users can easily update.
We can/should remove this in v9.

Includes https://github.com/getsentry/sentry-javascript/pull/11626
Fixes 49946