test(feedback): Re-add feedback CDN tests (#11890)
This reverts https://github.com/getsentry/sentry-javascript/pull/11888,
and ensures the feedback tests actually work on CDN.

For this, we now ensure to serve this locally, so this will work also on
release branches. It means you have to use `getLocalTestUrl` instead of
`getLocalTestPath` to work. (side note: We can/should probably just
remove `getLocalTestPath` overall 🤔 URL based is much more realistic and
IMHO better. I'll do that in a follow up, maybe.