feat(aws-serverless): Fix tree-shaking for aws-serverless package (#12017)
The prior fix was incomplete, because we were still using
`getDefaultIntegrations()` inside of node's `init`, so the deps where
pulled in anyhow.

Furthermore, it seems that `preserveModules: false` for `@sentry/node`
also prevented this from working as expected.

So this PR does two things:

1. Set `preserveModules: true` so that tree-shaking can work as expected
(😿 )
2. Expose a new `initWithoutDefaultIntegrations` method from
`@sentry/node` which AWS uses, which avoids including any integrations
by default. You have to pass your own `defaultIntegrations` to it.

This is not the prettiest solution, but I couldn't think of anything
much better 😬 I also added a size-limit entry to keep track of this.