test(e2e): Update node version used for e2e tests (#12031)
Extracted this out from
https://github.com/getsentry/sentry-javascript/pull/12020

For whatever reason, this does not seemed to play well with how we run
the proxy server via ts-node. Looking into this more, this is actually
not really necessary at all, so I just stripped this out in favor of
just running the proxy as a `.mjs` file directly - works just as well!