fix(core): Avoid looking up client for `hasTracingEnabled()` if possible (#12066)
This is a bit weird in that we access the current scope & client etc.,
even if options are passed. Since we also use this method in `init()`
before we have setup stuff, it seems safer to avoid calling this at all
if possible.

This is related to
https://github.com/getsentry/sentry-javascript/issues/12054, but not
really the cause of it.