chore: Align and update MIT license dates (#12143)
In https://github.com/getsentry/sentry-python/pull/3029 it was raised
that we need to make sure the SDK MIT License has an up-to-date year. We
also need to make it a range.

In this PR I went ahead and updated the dates accordingly, changed the
license to directly reference `Functional Software, Inc. dba Sentry`,
and re-aligned all the Licenses of individual packages to have the same
format and structure.