ref: Fix circular dependency checks & setup (#12159)
I noticed we were getting some build warnings for circular dependency
issues, which indicated that madge was not correctly capturing all
issues. So I went ahead and updated madge to latest, and aligned the
setup, which actually surfaced problems again more correctly.

I also fixed all the circular dep issues that were raised there. We
still have a bunch of types-related circular deps but these are hard to
solve, so let's ignore this for now.