fix(aws): Ensure lambda layer uses default export from `ImportInTheMiddle` (#12233)
Our lambda layer relies on one bundled Sentry SDK, which caused problems
raised in #12089 and
https://github.com/getsentry/sentry-javascript/issues/12009#issuecomment-2126211967.
Specifically, by bundling `import-in-the-middle` code into one file, it
seems like the library's way of declaring its exports conflict, causing
the "ImportInTheMiddle is not a constructor" error to be thrown. While
this should ideally be fixed soon in `import-in-the-middle`, for now
this patch adds a small Rollup plugin to transform `new
ImportInTheMiddle(...)` calls to `new ImportInTheMiddle.default(...)` to
our AWS Lambda Layer bundle config.