fix(browser): Improve browser extension error message check (#12146)
Make our check for when we abort and log an error due to
`Sentry.init` being used in a browser extension a bit more fine-gained.
In particular, we now do not abort the SDK initialization if we detect
that the SDK is running in a browser-extension dedicated window (e.g. a
URL starting with `chrome-extension://`).