fix(tracing): Ensure sent spans are limited to 1000 (#12252)
To avoid too large payloads, we should only send up to 1000 spans.

We actually had two different problems here in browser and node:

1. In browser, we _did_ limit to 1000, but we did so in an unfortunate
way, which is to reset the set after 1000 spans. So if a span had 1010
children, only the last 10 would be send.
2. In node, we just sent all spans.

I rewrote this to now consistently send the first 1000 spans of a
transaction.

I decided to keep all spans (no matter the limit) on the internal
parent-child map. Due to this this may grow a lot, but we can avoid
inconsistent state (e.g. what happens if a span has a parent in the map,
but the parent does not have the span as children, ...)
All of these should be garbage collected together ideally, so this
should be fine hopefully.