build: Bump node to 18.20.3
I noticed that locally the node-integration-tests that check versions fail, because in the SDK we use 18.18.0, we only check for "If node.major >= 18", and then get an incorrect test because actually in code we check the minor too.

Instead of adjust this, IMHO it's better to update this. This also never failed on CI because there we do use latest 18 for the matix tests...