fix(browser): Fix idle span ending (#12306)
I noticed here https://github.com/getsentry/sentry-javascript/pulls that
some tests were becoming flaky. Investigating this some more, I figured
out what was happening:

1. For idle spans, when the span is ended (in `on('spanEnded')`, which
is triggered in `span.end()`), we call `onIdleSpanEnded`
2. In there, we call `beforeSpanEnd`
3. This is used by `browserTracingIntegration` to call
`addPerformanceEntries(span)`, which adds performance spans to the idle
span
4. After that, in `onIdleSpanEnded`, any spans having start/end
timestamps outside of the idle span start/end timestamp will be
discarded

This lead to cases where performance spans were discarded because they
were out of bounds of the idle span - which is not what we want!

Now, this PR changes the timing of this a bit:

Now, we actually patch `span.end` of the SentrySpan, to ensure that we
can always run `beforeSpanEnd` with the correct timing, taking all spans
that are added in `beforeSpanEnd` into account to adjust the start/end
time of the idle span.