test: Refactor utils into `@sentry-internal/test-utils` (#12277)
Re-use this for playwright config in E2E tests.

Now, instead of repeating the whole playwright config, we can import the
util from `@sentry-internal/test-utils` and pass it some config.

I left the places where do not yet use the proxy, because for simplicity
I built the util in a way that assumes the proxy (you can override it,
but I figured not worth it for stuff that we will refactor soon anyhow).

While at it, I also streamlined the playwright version used everywhere
to an up-to-date version.