test(aws-lambda): Add basic lambda layer e2e test (#12279)
Add an e2e (or rather integration) test for our AWS lambda
layer bundle. The motivation for this test is that we broke the layer
during the initial v8 releases (multiple times for different reasons)
without us noticing this in tests. Simply because we never tested the
bundled SDK code that we create and publish for the lambda layer.