fix(browser): Fix INP span creation & transaction tagging (#12372)
Instead of https://github.com/getsentry/sentry-javascript/pull/12358,
this is a simpler change which ensures we pick the transaction from the
scope instead.

I also added tests for the various different scenarios, to ensure we see
how they behave:

1. INP is emitted _during_ pageload (span is active)
2. INP is emitted _after_ pageload

a. Pageload is parametrized (route)
b. Pageload is unparametrized (URL)

When the pageload is unparametrized (default browser SDK), the
transaction is not added to the DSC envelope header (which is correct
and also what we do in other places). it is _always_ added to the span
attributes now, though. If no span is active, it will use
transactionName from the last active pageload/navigation span.

There may be edge cases where this is not 100% correct (e.g. if the INP
span is only emitted once the pageload is done but another navigation
already started) but IMHO these are more edge cases and this change is
probably fine for now..?

(While at it, I also added an origin to the INP spans)