ref(node) refactor contextlines to use readline (#12221)
Refactors contextlines integration to use readline instead of reading
the entire file into memory (we were also doubling the memory as we were
splitting the file via `.split(\n)`.

Current implementation is not leveraging the cache to optimize file
reads, so I need to reintroduce that. Something to note is that the
caching behavior will become per frame location vs per file, which I
think makes more sense as we can only assume that this stack frame might
throw in the future (as opposed to anything from the file it is in).