feat(core): Add `parentSpan` option to `startSpan*` APIs (#12567)
With this PR, you can now pass a `parentSpan` optionally to `startSpan*`
APIs to create a span as a child of a specific span:


```js
const span = Sentry.startInactiveSpan({ name: 'xxx', parentSpan: parent });

Sentry.startSpan({ name: 'xxx', parentSpan: parent }, () => {});

Sentry.startSpanManual({ name: 'xxx', parentSpan: parent }, () => {});
```

With this, it should be easier to understand how you can manually work
with spans in v8.

Closes https://github.com/getsentry/sentry-javascript/issues/12539

ref https://github.com/getsentry/sentry-javascript/issues/12566, which I
noticed while working on this. For now I just cast the span which should
be fine, but is not ideal.