fix(v7/bundle): Ensure CDN bundles do not overwrite `window.Sentry` (#12579)
This was brought up in Slack - turns out we were always overwriting
`window.Sentry` when a CDN bundle was loaded. This means that if a user
e.g. loaded an integration via CDN, which is put on `window.Sentry`,
this was overwritten if the loader fetched the base CDN bundle later.

This _probably_ did not come up earlier because we may handle
`Sentry.Integrations` specially, but not sure. This fix should ensure
this works as expected!