feat(build): Drop prepack step & directly pack `/build` directory (#12656)
This is mostly an internal change, but it does change the layout of the
published package, which shouldn't affect users, unless they depend on
non-public things (which nobody should!).

With this change, we will simply pack the packages as-is without moving
any folder relativity. This means we do not need a `prepack` step
anymore.

Instead, we define which files should be included in the tarball via
`files` in the `package.json`.

<img width="497" alt="image"
src="https://github.com/getsentry/sentry-javascript/assets/2411343/39aeba01-eff2-41b3-afc2-6e437a23be96">

Fixes https://github.com/getsentry/sentry-javascript/issues/12642

There are some tiny fluctuations in the comparison table but as far as I
can tell nothing serious/bad.