feat(opentelemetry): Expose sampling helper (#12674)
When users want to use a custom sampler, they can use these new helpers
to still have sentry working nicely with whatever they decide to do in
there.

For e.g. trace propagation etc. to work correctly with Sentry, we need
to attach some things to trace state etc. These helpers encapsulate this
for the user, while still allowing them to decide however they want if
the span should be sampled or not.

This was brought up here:
https://github.com/getsentry/sentry-javascript/discussions/12191#discussioncomment-9892777