feat(core): allow unregistering callback through `on` (#11710)
This commit updates the return signature of the client's `on` function
to be a void function,
which, when executed, unregisters a callback. This adjustment is
necessary for managing instances
where objects are created and destroyed, ensuring that callbacks are
properly unregistered to
prevent self-referencing in callback closures and facilitate proper
garbage collection.

Typically, changing a type from `void` to `() => void` (or
`VoidFunction`) shouldn't be considered
a breaking change because `void` signifies the absence of a return
value, implying that the return
value of the `on` function should never be used by consumers.

Opting for the `on` approach, which returns a cleanup function, "seems"
simpler because having another
function called `off` requires saving the callback reference for later
removal. With our pattern,
we encapsulate both the registration and removal of event listeners
within a single function call.

---------

Co-authored-by: Francesco Novy <francesco.novy@sentry.io>