fix(tracing): Report dropped spans for transactions (#12751)
We already report a client report for dropped standalone spans. However,
so far we did not report any for spans dropped for transactions.

This PR also emits a client report when `beforeSendSpan` is used to
filter a span of a transaction, as well as when the whole transaction is
dropped in `beforeSendTransaction`.

For now, we _do not_ emit this if users manually drop single spans in
e.g. `beforeSendTransaction`, which I'd say is OK for now.

Closes https://github.com/getsentry/sentry-javascript/issues/12727