test(remix): Add dependency resolution overrides to fix breaking Node compatibility changes in dependencies (#12793)
This PR fixes our broken Remix v1 @ Node 16 integration test. Over the
weekend, the following transitive dependencies received minor and patch
releases which removed support for EOL Node versions.

- `glob` (https://github.com/isaacs/node-glob/issues/596)
- `lru-cache` (https://github.com/isaacs/node-lru-cache/issues/340)
- `jackspeack` (https://github.com/isaacs/jackspeak/issues/13)

IMO this, despite the dropped versions being EOL, is a breaking change
and a major inconvenience for a lot of users. I added comments/opened
issues with the request to revert the Node version drops and do it in a
major version instead. We're by far not the only affected ones. As our
tests show, at least everyone using Remix@1 on Node 16 will be broken by
this.

To fix, I added dependency resolutions, mostly scoped to specific
packages that depend on one of the three packages. Global overrides for
`glob` and `lru-cache` did not work because we have multiple
dependencies depending on different major versions of said packages.