fix(inp): Ensure INP spans have correct transaction (#12871)
Fixes https://github.com/getsentry/sentry-javascript/issues/12855

Previously, we stored the route name of the pageload in a map for INP
interactions. However, in some frameworks - e.g. remix, but also others,
we update the pageload span name later (since we rely on e.g. react
hooks for this etc).

Since we store the name of the pageload span at the time the first
interaction is recorded, it can thus happen that these run out of sync.

This PR changes this so that instead of the routename itself, we store
the pageload span in a map, and pick the last name of this when
generating the INP span.

I added tests in a remix e2e tests that show the now correct behavior,
these used to fail (because `transaction` on the pageload INP was `/`).