ref(e2e-tests): Make `startCommand` optional in shared playwright config (#12842)
Not all our e2e tests require a `startCommand` as we sometimes directly
invoke an application from within tests. For example in our AWS lambda
e2e tests. This PR makes the `startCommand` option optional so that we
can avoid a larger override of the entire `webServer` playwright config
object.