fix(node): Ensure correct URL is passed to `ignoreIncomingRequests` callback (#12929)
Fix an oversight in our Node `httpIntegration`. It looks like
we assumed that the `request` object being passed to
`ignoreIncomingRequestHook` and `ignoreOutgoingRequestHook` was of the
same type. However, it's not:
- `request` is of type `IncomingMessage` in `ignoreIncomingRequestHook`
- `request` is of type `RequestOptions` in `ignoreOutgoingRequestHook` 

fix the bug by simply taking the request.url property instead and adds integration tests to properly test the two options.