chore: Update changed package paths (#12954)
I noticed these while working on the cloudflare sdk.

- update browser path to use new replay packages

We were previously referencing `packages/replay`, which does not exist
anymore. This updates us to use the correct paths.

- add vercel-edge to nextjs

The nextjs sdk depends on vercel-edge, so we should be running tests for
it when that package changes.

- remove browser from deno

The deno sdk no longer relies on the browser sdk, this updates the paths
accordingly

- add bun section

We can be smarter about CI time and only run bun tests when relevant
package paths change.