feat(cloudflare): Add cloudflare sdk scaffolding (#12953)
This PR adds basic scaffolding for the cloudflare workers SDK. Most of
this is based on `@sentry/vercel-edge`.

This adds:

1. A basic cloudflare workers client
2. A set of default integrations for the cloudflare sdk (including a
fetch based one)
3. A cloudflare transport that uses the vercel-edge transport
4. An async context strategy powered by AsyncLocalStorage

You'll notice that there is no `init` method defined or exported from
the SDK. This is on purpose! `init` for cloudflare workers will work a
bit differently than the other SDKs, so I wanted to address it
differently on purpose. You'll see what that looks like in the next PR!