test(sveltekit): Switch to explicit vitest imports (#12960)
As per https://vitest.dev/config/#globals

> By default, vitest does not provide global APIs for explicitness

I think we should follow vitest defaults here and explicitly import in
the APIs that we need. This refactors our SvelteKit SDK tests to do so.

This change also removes `environment: 'jsdom'` from the sveltekit
global vite config in favour of explicitly adding jsdom environment via
the `@vitest-environment` pragma to the specific test file that needs
it. This should means that our server tests are not polluted with jsdom
globals, and that future writers have to explicitly opt-in to the
behaviour.