fix(node): Pass inferred name & attributes to `tracesSampler` (#12945)
Previously, we passed the span name directly to the `tracesSampler`
as-is. However, this is not ideal because this does not match the name
we actually send to sentry later, making this confusing. The reason is
that we infer the name from the attributes for some types of spans, but
we do that at export-time.

This PR adjust this so that we send the inferred name & attributes to
the `tracesSampler`. This works reasonably enough. However, there is a
big caveat: This still only has access to the initial attributes that a
span has at creation time. This means that we'll never have e.g. the
`http.route` correctly there, because all `http.server` spans originate
from the http instrumentation where they do not have a route yet, and
then more specific instrumentation (e.g. express or fastify) add the
`http.route` to that span later. So the name will never be parametrized,
for example, for `http.server` spans, which is not ideal (as it will
still not match the final span exactly) but should still be better than
what we had so far (where the name would always just be e.g. `GET`).

Fixes https://github.com/getsentry/sentry-javascript/issues/12944